feat: finished manager code, implemented manager api, restructured project directory and implemented socket communication among containers

+++ monitor/workers/agent.py
@@ -4,10 +4,15 @@ import os
 executing_dir, _ = os.path.split(os.path.abspath(__file__))
 sys.path.append(os.path.join(executing_dir, '..'))
 
-from models.agent import Agent
+from services.agent import Agent
+from errors.errors import EnvVariableMissing
 
 def main():
     metrics_path = os.getenv('METRICS_PATH')
+    agent_id = os.getenv('AGENT_ID')
+
+    if not agent_id:
+        raise EnvVariableMissing('AGENT_ID environment variable is missing')
 
     if metrics_path and os.path.exists(metrics_path):
         os.remove(metrics_path)
